djangoblog/requirements.txt
aliyun-python-sdk-core==2.13.31
aliyun-python-sdk-kms==2.14.0
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
Markdown==3.3.4
oss2==2.14.0
Pillow==8.1.2
pycparser==2.20
pycryptodome==3.10.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.4

selenium~=3.141.0
django-debug-toolbar~=3.2